+++ src/picoc/heap.c
@@ -29,7 +29,7 @@ void HeapInit(Picoc *pc, int StackOrHeapSize)
     pc->StackFrame = NULL;  /* the current stack frame */
     pc->HeapStackTop = NULL;  /* the top of the stack */
 
-    while (((unsigned long)&pc->HeapMemory[AlignOffset] & (sizeof(ALIGN_TYPE)-1)) != 0)
+    while (((uintptr_t)&pc->HeapMemory[AlignOffset] & (sizeof(ALIGN_TYPE)-1)) != 0)
         AlignOffset++;
 
     pc->StackFrame = &(pc->HeapMemory)[AlignOffset];
